backend/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2
APScheduler==3.6.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
Automat==24.8.1
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
minio==7.2.10
netifaces==0.10.6
openpyxl==3.1.5
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.3.0
python-multipart==0.0.12
python-socketio==5.11.4
pytz==2024.2
rsa==4.9
service-identity==24.2.0
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.39.2
tomli==2.0.2
Twisted==24.7.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.31.1
wsproto==1.2.0
zope.interface==7.1.0
